+++ examples/data-sources/ceph_pool/data-source.tf
@@ -1,0 +1,7 @@
+data "ceph_pool" "rbd" {
+  name = "rbd"
+}
+
+output "rbd_pg_num" {
+  value = data.ceph_pool.rbd.pg_num
+}